+++ packages/vite/src/node/__tests__/config.spec.ts
@@ -1,7 +1,8 @@
 import { describe, expect, test } from 'vitest'
 import type { InlineConfig } from '..'
 import type { UserConfig, UserConfigExport } from '../config'
-import { resolveConfig, resolveEnvPrefix } from '../config'
+import { resolveConfig } from '../config'
+import { resolveEnvPrefix } from '../env'
 import { mergeConfig } from '../publicUtils'
 
 describe('mergeConfig', () => {